+++ NEWS.md
@@ -14,6 +14,9 @@
 - `image()`: support of size in cm or inches
 - `instructions()`: Added `subtitle`, `top_space` parameter and color control
 
+### Minor changes
+- Added "teal" color style for scales
+
 # CURRENT STABLE: 0.1.3 (2016-11-12)
 Pushed to pypi!
 ### New functions